--- mbl-discord-client-main/requirements.txt
@@ -1,2 +1,0 @@
-python-dotenv
-discord.py
